Module ic_messaging::certified_slice_pool

A pool of incoming CertifiedStreamSlices used by XNetPayloadBuilderImpl to build XNetPayloads without the need for I/O on the critical path.

Structs

CertifiedSlicePool
A pool of CertifiedStreamSlices that provides support for taking out valid certified sub-slices of arbitrary size starting from arbitrary stream indices; and maintains cached stream positions, so an asynchronous process may populate the pool in the background with appropriate slices.

UnpackedStreamSlice
An unpacked CertifiedStreamSlice: a slice of the stream of messages produced by a subnet together with a cryptographic proof that the majority of that subnet agrees on it.

Enums

CertifiedSliceError
Error type returned when a pool operation failed due to an invalid slice.

InvalidAppend
Root cause of append() failure.

InvalidSlice
CertifiedSliceError::InvalidPayload and CertifiedSliceError::InvalidWitness detail.

Constants

LABEL_STATUS

METRIC_POOL_SIZE_BYTES

METRIC_TAKE_COUNT

METRIC_TAKE_GCED_MESSAGES

METRIC_TAKE_MESSAGES

METRIC_TAKE_SIZE_BYTES

STATUS_NONE

STATUS_SUCCESS

Type Definitions

CertifiedSliceResult
Result type returned by CertifiedSlicePool operations.