Struct ic_messaging::MessageRoutingImpl

pub struct MessageRoutingImpl { /* fields omitted */ }

Implementation of the MessageRouting trait.

Implementations

impl MessageRoutingImpl

pub fn new(
state_manager: Arc<dyn StateManager<State = ReplicatedState>>,
certified_stream_store: Arc<dyn CertifiedStreamStore>,
ingress_history_writer: Arc<dyn IngressHistoryWriter<State = ReplicatedState> + 'static>,
exec_env: Arc<dyn ExecutionEnvironment<State = ReplicatedState, CanisterState = CanisterState> + 'static>,
cycles_account_manager: Arc<CyclesAccountManager>,
scheduler_config: SchedulerConfig,
subnet_id: SubnetId,
subnet_type: SubnetType,
metrics_registry: &MetricsRegistry,
log: ReplicaLogger,
registry: Arc<dyn RegistryClient>
) -> Self

Creates a new MessageRoutingImpl for the given subnet using the provided StateManager and ExecutionEnvironment.

pub fn new_fake(
subnet_id: SubnetId,
state_manager: Arc<dyn StateManager<State = ReplicatedState>>,
ingress_history_writer: Arc<dyn IngressHistoryWriter<State = ReplicatedState> + 'static>,
metrics_registry: &MetricsRegistry,
log: ReplicaLogger
) -> Self

Creates a new MessageRoutingImpl for the given subnet using a fake BatchProcessor and the provided StateManager.